Moved new plot functionality from script to `wizardhat.plot`, and added preliminary blocking functionality to update plot whenever new samples are added to `TimeSeries`. Added `last_samples` property to `TimeSeries` and renamed `last` property to `last_sample`, and updated reference in `wizardhat.acquire`. Added bokeh to requirements.txt.
